Fixing Naming Convention Mistake in User Seeder Function
The seed_users function in the seeder.py file had a mistake in the naming convention for the User model, which was using "Users" instead of "User". This caused an error when attempting to delete all users from the database. The mistake was fixed by changing "Users" to "User" in the appropriate line of code.